pyproject.toml
[project]
name = "daydream-whisper"
version = "0.2.0"
description = "Apple Silicon speech recognition CLI — powered by MLX Whisper"
requires-python = ">=3.14"
dependencies = [
    "mlx-whisper>=0.4.0",
    "click>=8.1",
    "huggingface-hub>=0.20",
    "pyyaml>=6.0",
    "rich>=13.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24",
]

[project.scripts]
dwhisper = "dwhisper.cli:cli"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["dwhisper*"]
